+++ CHANGELOG.md
@@ -5,12 +5,13 @@ this file.
 ## 1.8 - 2025/09/27
 ### Added
 * Added Chasca, Ororon, Citlali, Mavuika, Lan Yan, Yumemizuki Mizuki, Iansan, Varesa, Escoffier, Ifa, Dahlia, Skirk, Ineffa, Aino, Lauma
+* Added chat:read scope in the GUI app to make Twitch chat commands work
 ### Changed
 * Changed Promise-dependent JQuery code to vanilla JavaScript
 ### Fixed
 * WORKING AGAIN! Overhauled the backend: migrated from PubSub to EventSub
 ### Removed
-* Removed dependency on JQuery and parseIRC
+* Removed dependency on JQuery
 
 ## 1.7.1 - 2024/10/18
 ### Added